support replaceExisting parameter for segments pushers

+++ extensions-contrib/cloudfiles-extensions/src/main/java/io/druid/storage/cloudfiles/CloudFilesDataSegmentPusher.java
@@ -74,7 +74,8 @@ public class CloudFilesDataSegmentPusher implements DataSegmentPusher
   }
 
   @Override
-  public DataSegment push(final File indexFilesDir, final DataSegment inSegment) throws IOException
+  public DataSegment push(final File indexFilesDir, final DataSegment inSegment, final boolean replaceExisting)
+      throws IOException
   {
     final String segmentPath = CloudFilesUtils.buildCloudFilesPath(this.config.getBasePath(), getStorageDir(inSegment));
 
@@ -98,18 +99,23 @@ public class CloudFilesDataSegmentPusher implements DataSegmentPusher
                   segmentPath, outFile, objectApi.getRegion(),
                   objectApi.getContainer()
               );
-              log.info("Pushing %s.", segmentData.getPath());
-              objectApi.put(segmentData);
 
-              // Avoid using Guava in DataSegmentPushers because they might be used with very diverse Guava versions in
-              // runtime, and because Guava deletes methods over time, that causes incompatibilities.
-              Files.write(descFile.toPath(), jsonMapper.writeValueAsBytes(inSegment));
-              CloudFilesObject descriptorData = new CloudFilesObject(
-                  segmentPath, descFile,
-                  objectApi.getRegion(), objectApi.getContainer()
-              );
-              log.info("Pushing %s.", descriptorData.getPath());
-              objectApi.put(descriptorData);
+              if (!replaceExisting && objectApi.exists(segmentData.getPath())) {
+                log.info("Skipping push because object [%s] exists && replaceExisting == false", segmentData.getPath());
+              } else {
+                log.info("Pushing %s.", segmentData.getPath());
+                objectApi.put(segmentData);
+
+                // Avoid using Guava in DataSegmentPushers because they might be used with very diverse Guava versions in
+                // runtime, and because Guava deletes methods over time, that causes incompatibilities.
+                Files.write(descFile.toPath(), jsonMapper.writeValueAsBytes(inSegment));
+                CloudFilesObject descriptorData = new CloudFilesObject(
+                    segmentPath, descFile,
+                    objectApi.getRegion(), objectApi.getContainer()
+                );
+                log.info("Pushing %s.", descriptorData.getPath());
+                objectApi.put(descriptorData);
+              }
 
               final DataSegment outSegment = inSegment
                   .withSize(indexSize)
